// _js/snippets/product-options/app/modules/as-radios/style.module.scss.d.ts
/* DO NOT EDIT: This file is auto-generated and will be overwritten. [build-fingerprint:codegen] */
export type Styles = {
  "ProductRadios": "ProductRadios",
  "ProductRadios-group": "ProductRadios-group",
  "ProductRadios-groupTitle": "ProductRadios-groupTitle",
  "ProductRadios-groupTitle--focused": "ProductRadios-groupTitle--focused",
  "ProductRadios-item": "ProductRadios-item",
  "ProductRadios-item--outOfStock": "ProductRadios-item--outOfStock",
  "ProductRadios-item--unavailable": "ProductRadios-item--unavailable",
  "ProductRadios-itemInput": "ProductRadios-itemInput",
  "ProductRadios-itemLabel": "ProductRadios-itemLabel",
  "ProductRadios-list": "ProductRadios-list"
}

export type ClassNames = keyof Styles

declare const styles: Styles

export default styles